+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - 94 Jobs import HTML formatting improvements
 - 96 Jobs search: add category filter
 - 99 Add slugs to job category
+- 104 Add job subcategory
 - Update deployment documentation
 
 ## 3.2 (2020-02-17)